[Refactor]: AnonymousProfileImageRepository 패키지 변경
Related to: #559
