+++ src/main/java/org/sopt/makers/internal/community/repository/anonymous/AnonymousProfileImageRepository.java
